break out tradingview chart component
