+++ dashboard/.eslintcache
@@ -1,1 +1,1 @@
-[{"/home/rob/projects/scratch/yogo/yogo-dashboard/src/reportWebVitals.js":"1","/home/rob/projects/scratch/yogo/yogo-dashboard/src/index.jsx":"2","/home/rob/projects/scratch/yogo/yogo-dashboard/src/App.jsx":"3","/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Watchlist.jsx":"4","/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Tradevew.jsx":"5","/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Watch.jsx":"6"},{"size":362,"mtime":1611008313141,"results":"7","hashOfConfig":"8"},{"size":577,"mtime":1611171802782,"results":"9","hashOfConfig":"8"},{"size":495,"mtime":1611177858168,"results":"10","hashOfConfig":"8"},{"size":719,"mtime":1611180505161,"results":"11","hashOfConfig":"8"},{"size":804,"mtime":1611180052740,"results":"12","hashOfConfig":"8"},{"size":766,"mtime":1611177798544,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"sorxpf",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"/home/rob/projects/scratch/yogo/yogo-dashboard/src/reportWebVitals.js",[],["27","28","29","30","31"],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/index.jsx",[],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/App.jsx",[],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Watchlist.jsx",[],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Tradevew.jsx",[],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Watch.jsx",[],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},"lines-around-directive",["42"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements"]
+[{"/home/rob/projects/scratch/yogo/yogo-dashboard/src/reportWebVitals.js":"1","/home/rob/projects/scratch/yogo/yogo-dashboard/src/index.jsx":"2","/home/rob/projects/scratch/yogo/yogo-dashboard/src/App.jsx":"3","/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Watchlist.jsx":"4","/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Chart.jsx":"5","/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Watch.jsx":"6"},{"size":362,"mtime":1611008313141,"results":"7","hashOfConfig":"8"},{"size":577,"mtime":1611171802782,"results":"9","hashOfConfig":"8"},{"size":495,"mtime":1611177858168,"results":"10","hashOfConfig":"8"},{"size":719,"mtime":1611180505161,"results":"11","hashOfConfig":"8"},{"size":804,"mtime":1611180052740,"results":"12","hashOfConfig":"8"},{"size":766,"mtime":1611177798544,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"sorxpf",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"/home/rob/projects/scratch/yogo/yogo-dashboard/src/reportWebVitals.js",[],["27","28","29","30","31"],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/index.jsx",[],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/App.jsx",[],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Watchlist.jsx",[],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Chart.jsx",[],"/home/rob/projects/scratch/yogo/yogo-dashboard/src/watch/Watch.jsx",[],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},"lines-around-directive",["42"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements"]
