+++ dashboard/src/watch/Watch.jsx
@@ -5,7 +5,7 @@ import CardContent from "@material-ui/core/CardContent";
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
 import PropTypes from "prop-types";
-import Tradeview from "./Tradevew";
+import Chart from "../tradingview/Chart";
 
 export default function Watch({ symbol }) {
   return (
@@ -14,7 +14,7 @@ export default function Watch({ symbol }) {
         <Typography variant="h5" component="h2">
           {symbol}
         </Typography>
-        <Tradeview symbol={symbol} />
+        <Chart symbol={symbol} />
       </CardContent>
       <CardActions>
         <Button size="small">Learn More</Button>
